Treat errors as 0 for sums
I decided that I didn't want errors to destroy a sum. The use case is in
my finances program, where I don't want missing months to wipe out the
sum for the year.

I originally hacked together a flag to enable that, before deciding that
I want this to be the default behavior, after which I couldn't think of
a use case for the original behavior, and so wiped out the flag parsing
system.

I took this time to put error cases first, as is my convention these
days, while I was at it.